+++ tests/org.jboss.tools.hibernate.orm.test/META-INF/MANIFEST.MF
@@ -15,7 +15,8 @@ Require-Bundle: org.hibernate.eclipse,
  org.eclipse.jdt.core,
  org.eclipse.core.runtime;bundle-version="3.13.0",
  org.eclipse.jdt.ui,
- org.eclipse.ui.ide;bundle-version="3.13.0"
+ org.eclipse.ui.ide;bundle-version="3.13.0",
+ org.eclipse.text
 Bundle-ActivationPolicy: lazy
 Bundle-RequiredExecutionEnvironment: JavaSE-1.8
 Import-Package: org.eclipse.osgi.util
